chore: refactor pandal marker
